refactor: fix habits_display folder typo
