+++ lib/view/for_habits_display/habit_display_appbar_view_mode.dart
