+++ lib/view/page_habits_display.dart
@@ -54,8 +54,8 @@ import '_debug.dart';
 import 'common/_dialog.dart';
 import 'common/_mixin.dart';
 import 'common/_widget.dart';
-import 'for_habit_display/_dialog.dart';
-import 'for_habit_display/_widget.dart';
+import 'for_habits_display/_dialog.dart';
+import 'for_habits_display/_widget.dart';
 import 'page_app_setting.dart' as app_setting_view;
 import 'page_habit_detail.dart' as habit_detail_view;
 import 'page_habit_edit.dart' as habit_edit_view;
